// FrontMiJardin/src/app/interfaces/NotasCResponse.ts
export class NotasCResponse {
    idNotaContable: number;
    NombreCliente: string;
    userName: string;
    tipoNota: string;
    estado: string;
    fechaCreacion: string;
    monto: number;
    aprobadorNota: string;
    usuarioCreacion: string;
    mostrarRecall: boolean;
    idFlujoAprobacion: number;
    orden: number;
    observacion: string;
}