// FrontMiJardin/src/app/interfaces/nuevoProducto.ts
export class nuevoProducto {
    idProducto: string;
    idMoneda: string;
    moneda: string;
    cantidad: string;
    vlUnitario: string;
    unidadMedida: string;
    nLote: string;
    nFactura: string;
    eliminarProducto: string;
    nombreProducto: string;
    descripcion: string;

}